polsf_unet/pyproject.toml
[project]
name = "torchtmpl"
version = "0.0.1"

dependencies = [
    "PyYAML>=6.0",
    "torchinfo>=1.7.2",
    "wandb>=0.13.11",
    "matplotlib>=3.7.1",
    "torchcvnn>=0.9.4",
    "scikit-image>=0.22.0",
    "torchvision>=0.16",
    "scikit-learn>=1.0",
    "seaborn>=0.11.2",
    "tifffile>=2024.8.28",
    "Pillow>=10.0",
    "openpyxl>=3.0",
    "PyQt5>=5.15.10",
    "plotly>=5.24.0",
    "h5py>=3.11.0",
    "umap-learn>=0.5.2",
    "scipy>=1.10.0",
]

[options]
package_dir = "torchtmpl"
